refactor(crates): Move project directory setup functions to lib crate
