+++ crates/teamtype/src/lib.rs
@@ -12,6 +12,7 @@ pub mod config;
 pub mod daemon;
 pub mod logging;
 pub mod sandbox;
+pub mod setup;
 pub mod traits;
 pub mod types;
